4/28/16
At least this being a separate branch means that we don't have to worry about having this tangled mess of commits in our overall project history.
I got some work done before realizing that this isn't the most up to date version of the calendar page, so I get to look forward to untangling that later.

x/
r
t/

4/28/16

I apologize for spamming our history with a ridiculous number of commits, but at least it's not the main branch
I got some work done today, but as it turns out I forgot to commit my most recent version of the files, so I get to try to detangle them later. Hooray.